% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workshop2020.R
\docType{package}
\name{workshop2020-package}
\alias{workshop2020}
\alias{workshop2020-package}
\title{Basic and advanced bioinformatics for RegenerationNEXT}
\description{
With breakthroughs in Next Generation Sequencing (NGS), 
we are collecting vast amounts of data that will be analyzed for screening
interesting targets. The details of these data reveal basic information 
in genome-wide such as gene expression profiles, DNA-protein interactions,
DNA-DNA interactions and chromatin structures, or may lead us to major 
discoveries like gene-regeneration associations. This workshop is designed
for postdoc, graduated students to get the basic and advanced bioinformatics
skills used in RNA-seq and DNA-seq data analysis. We will focus on the 
fundamentals of sequence analysis and try to let every attendance run 
the pipeline with sample data in order to make researchers confident 
in sequencing data analysis, understanding the results and presenting the 
data by themselves.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jianhong.github.io/workshop2020}
  \item Report bugs at \url{https://github.com/jianhong/workshop2020/issues/new/choose}
}

}
